Rebalance sensor/jammer cost/range
Significantly more expensive because it is a hard counter and is mobile to cloak/stealth. The cost is based off omni + t3 pgen costs.
In return, the intel range is increased to better support the army: Omni 80 -> 100 (stand in front of fatboy to give max range), vision 45 -> 50 (slightly more vision to help Percivals from further away)